requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
weaviate-client==4.17.0
httpx==0.28.1
